refactor(koth): align fake filenames with the .fake.ts convention
