+++ pages/src/components/stats/score-progression/fakes/koth-hill-data.fake.ts
