+++ pages/src/components/stats/score-progression/fakes/score-progression.fake.ts
